requirements.txt
altair==4.0.1
altgraph==0.17
argh==0.26.2
astor==0.8.1
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==1.5.0
blinker==1.4
cachetools==3.1.1
certifi==2019.9.11
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
enum-compat==0.0.3
enum34==1.1.6
fake-useragent==0.1.11
future==0.18.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
Google-Search-API==1.1.14
html5lib==0.9999999
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.1.0
jupyter-core==4.6.3
keyboard==0.13.4
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
MouseInfo==0.1.2
multidict==4.5.2
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.17.3
oauthlib==3.1.0
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pathtools==0.1.2
pefile==2019.4.18
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
playsound==1.2.2
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.10.0
psutil==5.6.7
ptyprocess==0.6.0
py2exe==0.9.2.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAudio==0.2.11
PyAutoGUI==0.9.48
pycodestyle==2.5.0
pycparser==2.20
pyFirmata==1.1.0
PyGetWindow==0.0.8
pyglet==1.5.1
Pygments==2.6.1
PyInstaller==3.6
PyMsgBox==1.0.7
pynput==1.4.4
pyOpenSSL==19.1.0
pyperclip==1.7.0
pypiwin32==223
PyRect==0.1.4
pyrsistent==0.15.7
PyScreeze==0.1.25
pyserial==3.4
pysound==0.1
python-dateutil==2.8.0
python-utils==2.3.0
pyttsx3==2.71
PyTweening==1.0.3
pytz==2019.3
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==19.0.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
secure-smtplib==0.1.1
selenium==2.53.6
Send2Trash==1.5.0
serial==0.0.97
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
tcl-kernel==0.0.4
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
terminado==0.8.3
testpath==0.4.4
tflearn==0.3.2
timeloop==1.0.2
tkthread==0.3.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.0.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.6
vcrpy==2.1.0
watchdog==0.9.0
wcwidth==0.1.8
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wincertstore==0.2
wrapt==1.11.2
yarl==1.3.0
zipp==3.1.0
